requirements.txt
flask
flask-cors
rembg
tensorflow==2.17.1
pillow==11.0.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.6.0
opencv-python==4.10.0.84
matplotlib==3.8.0